[bpd] Fix ghistory/btb/jmp bug
The ghistory register was incorrectly being speculatively updated with "taken"
in a particular scenario -- if the BTB's prediction was overridden by the
backing BPD as not-taken, but a later JMP instruction in the packet existed, the
front-end would be correctly redirected to take the JMP by the BP2 stage, but
the BPD's ghistory register would be marked as "taken" as if the BTB had
predicted the taken branch. In fact, it should be marked as "not-taken", since
the ghistory does NOT contain JMP instructions, the branches in the fetch packet
were all predicted as not-taken, and the JMP is reason the front-end is being
redirected.