[bpd] get rid of shadow info.
The ghistory snapshot is NOT what ghistory was used to index the predictor,
rather it is the last copy of the history before the current branch
speculatively modified it. Getting rid of tracking this shadow information and
instead snapshotting a later copy of the history allows us to avoid an ugly
corner case.