Work in progress refactor to new code setup.
Previously, BOOM was set up as a fork of rocket-chip and hiding behind
BuildCore as a core-replacement over Rocket. This is problematic as BOOM may
want to provide modifications to the CoreIO (e.g., add PTW access for the
Load/Store Unit) or provide modifications to the caches.

This commit works towards using a new project-template repository setup in which
BOOM now creates a new BoomTile that can be mixed into rocket-chip systems and
in which BOOM depends on rocket-chip master (not a fork). This project-template
format more closely mirrors existing rocket-chip repositories and will hopefully
make it easier for users to port BOOM to new systems and platforms (e.g.,
FPGAs).

Currently a Work-in-Progress. VM tests do not work because the PTW ports are not
hooked up yet.

   * Now provide a BoomTile class that provides the same I/Os as a RocketTile,
      and inherits from BaseTile.
   * Provide a HasBoomTiles trait which can be mixed into rocket-chip systems.
   * Provide a BoomCoreplex which mixes in BoomTiles.
   * Provide an ExampleBoomSystem that instantiates a BoomCoreplex.
   * Provide a TestHarness which instantiates an ExampleBoomSystem.
   * Provide our own Generator so the BOOM repo can drive Chisel/FIRRTL.
   * The above changes required 500 loc to be duplicated from rocket-chip.
   * Added a system sub-directory to contain these new example system files
      (mimicking the rocket-chip organization).
   * BoomCoreParams now inherits from CoreParams instead of being a parallel
      off-shoot.
   * RocketCoreParams no longer used in BOOM.
   * BootROM path is changed since SBT will be called from a higherup directory.