[icache] Add ability to BlackBox the ICache.
Currently, to emit the BlackBox version of ICache, one must change the default
input parameter enableBlackBox in the ICache constructor.

BlackBoxing the ICache is somewhat challenging as the ICache is a LazyModule and
needs to do its job of negotiating and wiring stuff up. Likewise, it's also not
(currently) possible to BlackBox the LazyModuleImp, so the current solution is
to place the BlackBox inside of the LazyModuleImp, adding an extra level to the
hierarchy.

There is an additional challenge as the straight-forward solution leads to
firrtl-emitter crashes on a None.get. The current work-around is to add an extra
level of indirection around the IO bundle.