[dense-btb] restructed data arrays + fixed broken logic
* This commit changes the data arrays which were storing Vec of bundles
in a SeqMem to instead store bits and explicitly type-cast for when
an actual Vec is required. Without these changes, the generated verilog
for vsim was inferred correctly but the write mask signals were expanded
based on each field in the bundle. Need to revisit this after upgrading to
chisel3

* Fixed broken logic where not initializing a wire and not assigning the
wires conditionally resulted in inferred latches. Also, changed the
updates to the hits wire and the data_out wire based on the delayed pc request
to match the output of synchronous reads